add songsubmission component and refactor round page to include new component
